Label-row lanes hold a gene's transcripts past 255
The rect, line and arrow label-row lanes were Uint8, so in a gene with
more than 255 labelled transcripts under subfeatureLabels below the count
wrapped, and from the 257th isoform on the exons drew about 256 label rows
too high, over the first ones, while their hover boxes and labels stayed
put. The lanes are Uint16.

+++ plugins/canvas/src/LinearBasicDisplay/applyLayout.ts
@@ -20,7 +20,7 @@ function scaleFloat32(arr: Float32Array, multiplier: number) {
 // separately.
 function scaleYWithLabelRows(
   ys: Float32Array,
-  rows: Uint8Array,
+  rows: Uint16Array,
   multiplier: number,
   labelFontPx: number,
 ) {
